+++ tests/settings_tests.m
@@ -22,20 +22,29 @@ classdef settings_tests < matlab.unittest.TestCase
         function basicTest(obj)
             % Can we read in a current, correct, settings file?
             expected = obj.loadSettingsExample('zapitSystemSettings_01.mat');
-            actual = zapit.settings.readSettings(fullfile(obj.dataDir,'zapitSystemSettings_01.yml'));
+            actual = zapit.settings.readSettings(fullfile(obj.dataDir, ...
+                        'zapitSystemSettings_01.yml'));
             obj.verifyEqual(actual,expected)
         end
 
         function fixWrongValues(obj)
-            % Can we read in a current, correct, settings file?
+            % Can we fix incorrect values?
             expected = zapit.settings.default_settings;
-            actual = zapit.settings.readSettings(fullfile(obj.dataDir,'zapitSystemSettings_wrong_values_01.yml'));
+            actual = zapit.settings.readSettings(fullfile(obj.dataDir, ...
+                        'zapitSystemSettings_wrong_values_01.yml'));
 
             obj.verifyEqual(actual.general, expected.general)
             obj.verifyEqual(actual.NI, expected.NI)
-
         end
 
+        function checkRenameField(obj)
+            % Can we fix a field that has an old name to a new name?
+            expected = obj.loadSettingsExample('zapitSystemSettings_01.mat');
+            actual = zapit.settings.readSettings(fullfile(obj.dataDir, ...
+                        'zapitSystemSettings_fieldNameCHanged_01.yml'));
+
+            obj.verifyEqual(actual.experiment, expected.experiment)
+        end
     end %methods (Test)